swap over to ffi for globbing
