+++ cli/internal/doublestar/validate.go
@@ -3,27 +3,6 @@
 // SPDX-License-Identifier: MIT
 package doublestar
 
-import "path/filepath"
-
-// ValidatePattern validates a pattern. Patterns are validated while they run in Match(),
-// PathMatch(), and Glob(), so, you normally wouldn't need to call this.
-// However, there are cases where this might be useful: for example, if your
-// program allows a user to enter a pattern that you'll run at a later time,
-// you might want to validate it.
-//
-// ValidatePattern assumes your pattern uses '/' as the path separator.
-func ValidatePattern(s string) bool {
-	return doValidatePattern(s, '/')
-}
-
-// ValidatePathPattern only uses your OS path separator. In other words, use
-// ValidatePattern if you would normally use Match() or Glob(). Use
-// ValidatePathPattern if you would normally use PathMatch(). Keep in mind,
-// Glob() requires '/' separators, even if your OS uses something else.
-func ValidatePathPattern(s string) bool {
-	return doValidatePattern(s, filepath.Separator)
-}
-
 func doValidatePattern(s string, separator rune) bool {
 	altDepth := 0
 	l := len(s)
